Use double for clock times

+++ include/Game/Game.h
@@ -11,17 +11,17 @@ class Board;
 // Play a game with a new board in the default starting state.
 Color play_game(const Player& white,
                 const Player& black,
-                int time_in_seconds,
-                int moves_to_reset_clock,
-                int increment_seconds,
+                double time_in_seconds,
+                size_t moves_to_reset_clock,
+                double increment_seconds,
                 const std::string& pgn_file_name);
 
 // Play a game with a supplied board. Board's state may be modified.
 Color play_game_with_board(const Player& white,
                            const Player& black,
-                           int time_in_seconds,
-                           int moves_to_reset_clock,
-                           int increment_seconds,
+                           double time_in_seconds,
+                           size_t moves_to_reset_clock,
+                           double increment_seconds,
                            const std::string& pgn_file_name,
                            Board& board);
 #endif // GAME_H
